+++ emg3d/utils.py
@@ -1137,15 +1137,33 @@ class TensorMesh:
         return self.__vol
 
 
-def get_hx_h0(freq, rho, fixed, survey_domain, possible_nx, min_width=None,
-              pps=3, alpha=None, calc_domain_factors=None,
-              resp_survey_domain=False, raise_error=True, verb=1):
+def get_hx_h0(freq, res, domain, fixed=0., possible_nx=None, min_width=None,
+              pps=3, alpha=None, raise_error=True, verb=1, return_info=False):
     r"""Return cell widths and origin for given parameters.
 
-    Returns cell widths given the provided ``survey_domain`` and other
-    parameters using a flexible amount of cells. See input parameters for more
-    details. A maximum of three hard/fixed boundaries can be provided (one of
-    which is the grid center).
+    Returns cell widths for the provided frequency, resistivity, domain extent,
+    and other parameters using a flexible amount of cells. See input parameters
+    for more details. A maximum of three hard/fixed boundaries can be provided
+    (one of which is the grid center).
+
+    The minimum cell width is calculated through :math:`\delta/\rm{pps}`, where
+    the skin depth is given by :math:`\delta = 503.3 \sqrt{\rho/f}`, and
+    the parameter ``pps`` stands for 'points-per-skindepth'. The minimum cell
+    width can be restricted with the parameter ``min_width``.
+
+    The actual calculation domain adds a buffer zone around the (survey)
+    domain. The thickness of the buffer is six times the skin depth. The field
+    is basically zero after two wavelengths. A wavelength is
+    :math:`2\pi\delta`, hence roughly 6 times the skin depth. Taking a factor 6
+    gives therefore almost two wavelengths, as the field travels to the
+    boundary and back. The actual buffer thickness can be steered with the
+    ``res`` parameter.
+
+    One has to take into account that the air is very resistive, which has to
+    be considered not just in the vertical direction, but also in the
+    horizontal directions, as the airwave will bounce back from the sides
+    otherwise. In the marine case this issue reduces with increasing water
+    depth.
 
 
     See Also
@@ -1165,48 +1183,52 @@ def get_hx_h0(freq, rho, fixed, survey_domain, possible_nx, min_width=None,
         multiplied by :math:`-2\pi`, to simulate the closest
         frequency-equivalent.
 
-    rho : float
-        Resistivity (Ohm m) to calculate skin depth.
+    res : float or list
+        Resistivity (Ohm m) to calculate the skin depth. The skin depth is
+        used to calculate the minimum cell width and the boundary thicknesses.
+        Up to three resistivities can be provided:
 
-    fixed : array
+        - float: Same resistivity for everything;
+        - [min_width, boundaries];
+        - [min_width, left boundary, right boundary].
+
+    domain : list
+        Contains the survey-domain limits [min, max]. The actual calculation
+        domain consists of this domain plus a buffer zone around it, which
+        depends on frequency and resistivity.
+
+    fixed : list, optional
         Fixed boundaries, one, two, or maximum three values. The grid is
-        centered around the first value. Hence it is the location of the
-        smallest cell. Two more fixed boundaries can be added, at most one on
-        each side of the first one.
+        centered around the first value. Hence it is the center location with
+        the smallest cell. Two more fixed boundaries can be added, at most one
+        on each side of the first one.
+        Default is 0.
 
-    survey_domain : list {[min, max] or [min, seafloor, sea-surface]}
-        Contains the survey-domain limits. If the list contains three values,
-        then it is assumed to be the vertical direction, where the second and
-        third values are taken as seafloor and sea-surface. In this case, the
-        source has to be within the water layer.
+    possible_nx : list, optional
+        List of possible numbers of cells. See :func:`get_cell_numbers`.
+        Default is ``get_cell_numbers(500, 5, 3)``, which corresponds to
+        [16, 24, 32, 40, 48, 64, 80, 96, 128, 160, 192, 256, 320, 384].
 
-    possible_nx : list
-        List of possible number of cells. See :func:`get_cell_numbers`.
+    min_width : float, list or None, optional
+        Minimum cell width restriction:
 
-    min_width : list or None, optional
-        Minimum cell width restriction, [min, max].
+        - None : No restriction;
+        - float : Fixed to this value, ignoring skin depth and ``pps``.
+        - list [min, max] : Lower and upper bounds.
+
         Default is None.
 
     pps : int, optional
         Points per skindepth; minimum cell width is calculated via
-        `dmin = skind/pps`.
+        `dmin = skindepth/pps`.
         Default = 3.
 
     alpha : list, optional
-        Maximum alpha and how many steps it takes to find a good alpha. The
-        higher this number is, the longer it will take, but the finer it
-        samples alpha: [max. survey alpha, nr, max calc alpha, nr].
-        Default = [1.05, 11, 1.5, 11].
-
-    calc_domain_factors : list, optional
-        Defines boundary of how many skin-depths for
-        [min left, max left, min right, max right].
-        Default = [5, 10, 5, 10]
-
-    resp_survey_domain : bool, optional
-        If True, the survey_domain is respected even if the calculation_domain
-        is smaller.
-        Default is False.
+        Maximum alpha and step size to find a good alpha. The first value is
+        the maximum alpha of the survey domain, the second value is the maximum
+        alpha for the buffer zone, and the third value is the step size.
+        Default = [1, 1.5, .01], hence no stretching within the survey domain
+        and a maximum stretching of 1.5 in the buffer zone; step size is 0.01.
 
     raise_error : bool, optional
         If True, an error is raised if no suitable grid is found. Otherwise it
@@ -1217,6 +1239,10 @@ def get_hx_h0(freq, rho, fixed, survey_domain, possible_nx, min_width=None,
         Verbosity, 0 or 1.
         Default = 1.
 
+    return_info : bool
+        If True, a dictionary is returned with some grid info (min and max
+        cell width and alpha).
+
 
     Returns
     -------
@@ -1224,80 +1250,73 @@ def get_hx_h0(freq, rho, fixed, survey_domain, possible_nx, min_width=None,
         Cell widths of mesh.
 
     x0 : float
-        Origin corresponding to ``hx``.
+        Origin of the mesh.
 
     info : dict
-        Mesh information.
+        Dictionary with mesh info; only if ``return_info=True``.
+
+        Keys:
+
+        - ``dmin``: Minimum cell width;
+        - ``dmax``: Maximum cell width;
+        - ``amin``: Minimum alpha;
+        - ``amax``: Maximum alpha.
 
     """
     # Get variables with default lists:
-    alpha = alpha or [1.05, 11, 1.5, 11]
-    calc_domain_factors = calc_domain_factors or [5, 10, 5, 10]
+    alpha = alpha or [1, 1.5, 0.01]
+    possible_nx = possible_nx or get_cell_numbers(500, 5, 3)
 
-    # Start log.
-    if verb > 0:
-        space = 4*" "
-        print(f"  = = = = = = = = {freq:8.4f} Hz = = = = = = = =")
+    # Cast resistivity value(s).
+    res = np.array(res, ndmin=1)
+    if res.size == 1:
+        res_arr = np.array([res[0], res[0], res[0]])
+    elif res.size == 2:
+        res_arr = np.array([res[0], res[1], res[1]])
+    else:
+        res_arr = np.array([res[0], res[1], res[2]])
 
-    # Cast fixed.
+    # Cast and check fixed.
     fixed = np.array(fixed, ndmin=1)
     if fixed.size > 2:
-        if np.sign(np.diff(fixed[:2])) == np.sign(np.diff(fixed[::2])):
-            print("\n* ERROR   :: 2nd and 3rd fixed boundaries have to be "
-                  "left and right of the first one.\n             "
-                  f"Provided: [{fixed[0]}, {fixed[1]}, {fixed[2]}]")
-            raise ValueError("Wrong input for fixed")
+
+        # Check length.
         if fixed.size > 3:
             print("\n* ERROR   :: Maximum three fixed boundaries permitted.\n"
                   f"             Provided: {fixed.size}.")
             raise ValueError("Wrong input for fixed")
 
+        # Sort second and third, so it doesn't matter how it was provided.
+        fixed = np.array([fixed[0], max(fixed[1:]), min(fixed[1:])])
+
+        # Check side.
+        if np.sign(np.diff(fixed[:2])) == np.sign(np.diff(fixed[::2])):
+            print("\n* ERROR   :: 2nd and 3rd fixed boundaries have to be "
+                  "left and right of the first one.\n             "
+                  f"Provided: [{fixed[0]}, {fixed[1]}, {fixed[2]}]")
+            raise ValueError("Wrong input for fixed")
+
     # Calculate skin depth.
-    skind = 503.3*np.sqrt(rho/abs(freq))
+    skind = 503.3*np.sqrt(res_arr/abs(freq))
     if freq < 0:  # For Laplace-domain calculations.
         skind /= np.sqrt(2*np.pi)
-    if verb > 0:
-        print(space+f"Skin depth          [m] : {skind:.0f}")
 
     # Minimum cell width.
-    dmin = skind/pps
+    dmin = skind[0]/pps
     if min_width is not None:  # Respect user input.
-        dmin = np.clip(dmin, *np.array(min_width, dtype=float))
+        min_width = np.array(min_width, ndmin=1)
+        if min_width.size == 1:
+            dmin = min_width
+        else:
+            dmin = np.clip(dmin, *min_width)
 
     # Survey domain.
-    survey_domain = np.array(survey_domain, dtype=float)
-    if verb > 0:  # Print original survey domain info.
-        print(space+f"Survey domain       [m] : {survey_domain[0]:.0f} - "
-              f"{survey_domain[-1]:.0f}")
+    domain = np.array(domain, dtype=float)
 
     # Calculation domain.
-
-    # Get boundary.
-    calc_domain = skind*np.array(calc_domain_factors, dtype=float)
-
-    # Now get calculation domain by clipping survey to boundary.
-    calcmin = np.clip(survey_domain[0], *(fixed[0]-calc_domain[:2][::-1]))
-    calcmax = np.clip(survey_domain[-1], *(fixed[0]+calc_domain[2:]))
-    calc_domain = [calcmin, calcmax]
-    if verb > 0:  # Print original calculation domain info.
-        print(space+f"Calculation domain  [m] : {calc_domain[0]:.0f} - "
-              f"{calc_domain[1]:.0f}")
-
-    # If calc_domain < survey_domain, we restrict the calc_domain if
-    # resp_survey_domain is True, else the survey_domain.
-    if resp_survey_domain:
-        calc_domain[0] = min(survey_domain[0], calc_domain[0])
-        calc_domain[1] = max(survey_domain[-1], calc_domain[1])
-    else:
-        survey_domain[0] = max(survey_domain[0], calc_domain[0])
-        survey_domain[-1] = min(survey_domain[-1], calc_domain[1])
-
-    # Flag if there is no need to check the calc_domain.
-    if (np.isclose(calc_domain[0], survey_domain[0]) and
-            np.isclose(calc_domain[1], survey_domain[-1])):
-        same_bounds = True
-    else:
-        same_bounds = False
+    calc_domain = skind[1:]*np.array([6., 6.])  # 6 x sd => buffer zone.
+    calc_domain[0] = domain[0] - calc_domain[0]
+    calc_domain[1] = domain[1] + calc_domain[1]
 
     # Initiate flag if terminated.
     finished = False
@@ -1308,12 +1327,12 @@ def get_hx_h0(freq, rho, fixed, survey_domain, possible_nx, min_width=None,
     # Loop over possible cell numbers from small to big.
     for nx in np.unique(possible_nx):
 
-        # Loop over possible alphas for survey_domain.
-        for sa in np.linspace(1.0, alpha[0], alpha[1]):
+        # Loop over possible alphas for domain.
+        for sa in np.arange(1.0, alpha[0]+alpha[2]/2, alpha[2]):
 
             # Get current stretched grid cell sizes.
             thxl = dmin*sa**np.arange(nx)  # Left of origin.
-            thxr = dmin*sa**np.arange(nx)  # right of origin.
+            thxr = dmin*sa**np.arange(nx)  # Right of origin.
 
             # 0. Adjust stretching for fixed boundaries.
             if fixed.size > 1:  # Move mesh to first fixed boundary.
@@ -1326,27 +1345,25 @@ def get_hx_h0(freq, rho, fixed, survey_domain, possible_nx, min_width=None,
                 ii = np.argmin(abs(t_nx-fixed[2]))
                 thxl *= abs(fixed[2]-fixed[0])/np.sum(thxl[:ii])
 
-            # 1. Fill from center to left survey_domain.
-            nl = np.sum((fixed[0]-np.cumsum(thxl)) > survey_domain[0])+1
+            # 1. Fill from center to left domain.
+            nl = np.sum((fixed[0]-np.cumsum(thxl)) > domain[0])+1
 
-            # 2. Fill from center to right survey_domain.
-            nr = np.sum((fixed[0]+np.cumsum(thxr)) < survey_domain[-1])+1
+            # 2. Fill from center to right domain.
+            nr = np.sum((fixed[0]+np.cumsum(thxr)) < domain[1])+1
 
             # 3. Get remaining number of cells and check termination criteria.
-            nsdc = nl+nr  # Number of survey_domain cells.
+            nsdc = nl+nr  # Number of domain cells.
             nx_remain = nx-nsdc
-            if same_bounds and nx_remain >= 0:
-                # Good enough, distribute remaining cells.
-                nl += int(np.floor(nx_remain/2))  # If uneven, add one cell
-                nr += int(np.ceil(nx_remain/2))   # more on the right.
-            elif nx_remain <= 0:  # Not good, try next.
+
+            # Not good, try next.
+            if nx_remain <= 0:
                 continue
 
             # Create the current hx-array.
             hx = np.r_[thxl[:nl][::-1], thxr[:nr]]
             hxo = np.r_[thxl[:nl][::-1], thxr[:nr]]
 
-            # Get actual survey_domain:
+            # Get actual domain:
             asurv_domain = [fixed[0]-np.sum(thxl[:nl]),
                             fixed[0]+np.sum(thxr[:nr])]
             x0 = float(fixed[0]-np.sum(thxl[:nl]))
@@ -1354,13 +1371,8 @@ def get_hx_h0(freq, rho, fixed, survey_domain, possible_nx, min_width=None,
             # Get actual stretching (differs in case of fixed layers).
             sa_adj = np.max([hx[1:]/hx[:-1], hx[:-1]/hx[1:]])
 
-            # If good enough, mark it as finished and break out of the loop.
-            if same_bounds and nx_remain >= 0:
-                finished = True
-                break
-
             # Loop over possible alphas for calc_domain.
-            for ca in np.linspace(sa, alpha[2], alpha[3]):
+            for ca in np.arange(sa, alpha[1]+alpha[2]/2, alpha[2]):
 
                 # 4. Fill to left calc_domain.
                 thxl = hx[0]*ca**np.arange(1, nx_remain+1)
@@ -1408,33 +1420,44 @@ def get_hx_h0(freq, rho, fixed, survey_domain, possible_nx, min_width=None,
             hx, x0 = None, None
 
     elif verb > 0:
-        print(space+f"Final extent        [m] : {x0:.0f} - "
+        print(f"   Skin depth ", end="")
+        if res.size == 1:
+            print(f"         [m] : {skind[0]:.0f}")
+        elif res.size == 2:
+            print(f"(m/l-r)  [m] : {skind[0]:.0f} / {skind[1]:.0f}")
+        else:
+            print(f"(m/l/r)  [m] : {skind[0]:.0f} / {skind[1]:.0f} / "
+                  f"{skind[2]:.0f}")
+        print(f"   Survey domain       [m] : {domain[0]:.0f} - "
+              f"{domain[1]:.0f}")
+        print(f"   Calculation domain  [m] : {calc_domain[0]:.0f} - "
+              f"{calc_domain[1]:.0f}")
+        print(f"   Final extent        [m] : {x0:.0f} - "
               f"{x0+np.sum(hx):.0f}")
-        extstr = space+f"Min/max cell width  [m] : {min(hx):.0f} / "
-        alstr = space+f"Alpha survey"
-        nrstr = space+"Number of cells "
+        extstr = f"   Min/max cell width  [m] : {min(hx):.0f} / "
+        alstr = f"   Alpha survey"
+        nrstr = "   Number of cells "
         if not np.isclose(sa, sa_adj):
             sastr = f"{sa:.3f} ({sa_adj:.3f})"
         else:
             sastr = f"{sa:.3f}"
-        if same_bounds:
-            print(extstr+f"{max(hx):.0f}")
-            print(alstr+f"            : {sastr}")
-            print(nrstr+f"  (s/r) : {nx} ({nsdc}/{nx_remain})")
-        else:
-            print(extstr+f"{max(hxo):.0f} / {max(hx):.0f}")
-            print(alstr+f"/calc       : {sastr} / {ca:.3f}")
-            print(nrstr+f"(s/c/r) : {nx} ({nsdc}/{ncdc}/{nx_remain2})")
+        print(extstr+f"{max(hxo):.0f} / {max(hx):.0f}")
+        print(alstr+f"/calc       : {sastr} / {ca:.3f}")
+        print(nrstr+f"(s/c/r) : {nx} ({nsdc}/{ncdc}/{nx_remain2})")
         print()
 
-    if not fixed.size > 1:
-        sa_adj = sa
-    info = {'dmin': dmin,
-            'dmax': np.nanmax(hx),
-            'amin': np.nanmin([ca, sa, sa_adj]),
-            'amax': np.nanmax([ca, sa, sa_adj])}
+    if return_info:
+        if not fixed.size > 1:
+            sa_adj = sa
 
-    return hx, x0, info
+        info = {'dmin': dmin,
+                'dmax': np.nanmax(hx),
+                'amin': np.nanmin([ca, sa, sa_adj]),
+                'amax': np.nanmax([ca, sa, sa_adj])}
+
+        return hx, x0, info
+    else:
+        return hx, x0
 
 
 def get_cell_numbers(max_nr, max_prime=5, min_div=3):
@@ -1493,11 +1516,11 @@ def get_cell_numbers(max_nr, max_prime=5, min_div=3):
     return numbers[numbers <= max_nr]
 
 
-def get_stretched_h(h_min, domain, nx, x0=0, x1=None, resp_domain=False):
+def get_stretched_h(min_width, domain, nx, x0=0, x1=None, resp_domain=False):
     """Return cell widths for a stretched grid within the domain.
 
     Returns ``nx`` cell widths within ``domain``, where the minimum cell width
-    is ``h_min``. The cells are not stretched within ``x0`` and ``x1``, and
+    is ``min_width``. The cells are not stretched within ``x0`` and ``x1``, and
     outside uses a power-law stretching. The actual stretching factor and the
     number of cells left and right of ``x0`` and ``x1`` are find in a
     minimization process.
@@ -1522,9 +1545,9 @@ def get_stretched_h(h_min, domain, nx, x0=0, x1=None, resp_domain=False):
     Parameters
     ----------
 
-    h_min : float
+    min_width : float
         Minimum cell width. If x1 is provided, the actual minimum cell width
-        might be smaller than h_min.
+        might be smaller than min_width.
 
     domain : list
         [start, end] of model domain.
@@ -1540,7 +1563,7 @@ def get_stretched_h(h_min, domain, nx, x0=0, x1=None, resp_domain=False):
         If provided, then no stretching is applied between ``x0`` and ``x1``.
         The non-stretched part starts at ``x0`` and stops at the first possible
         location at or after ``x1``. ``x1`` is restricted to ``domain``. This
-        will h_min so that an integer number of cells fit within x0 and x1.
+        will min_width so that an integer number of cells fit within x0 and x1.
 
     resp_domain : bool
         If False (default), then the domain-end might shift slightly to assure
@@ -1562,7 +1585,7 @@ def get_stretched_h(h_min, domain, nx, x0=0, x1=None, resp_domain=False):
     domain = np.array(domain, dtype=float)
     x0 = np.array(x0, dtype=float)
     x0 = np.clip(x0, *domain)  # Restrict to model domain
-    h_min = np.array(h_min, dtype=float)
+    min_width = np.array(min_width, dtype=float)
     if x1 is not None:
         x1 = np.array(x1, dtype=float)
         x1 = np.clip(x1, *domain)  # Restrict to model domain
@@ -1574,23 +1597,23 @@ def get_stretched_h(h_min, domain, nx, x0=0, x1=None, resp_domain=False):
         xlim_orig = domain.copy()
         nx_orig = int(nx)
         x0_orig = x0.copy()
-        h_min_orig = h_min.copy()
+        h_min_orig = min_width.copy()
 
         # Get number of non-stretched cells
-        n_nos = int(np.ceil((x1-x0)/h_min))
+        n_nos = int(np.ceil((x1-x0)/min_width))
 
-        # Re-calculate h_min to fit with x0-x1-limits:
-        h_min = (x1-x0)/n_nos
+        # Re-calculate min_width to fit with x0-x1-limits:
+        min_width = (x1-x0)/n_nos
 
         # Subtract one cell, because the standard scheme provides one
-        # h_min-cell.
+        # min_width-cell.
         n_nos -= 1
 
-        # Reset x0, because the first h_min comes from normal scheme
-        x0 += h_min
+        # Reset x0, because the first min_width comes from normal scheme
+        x0 += min_width
 
         # Reset xmax for normal scheme
-        domain[1] -= n_nos*h_min
+        domain[1] -= n_nos*min_width
 
         # Reset nx for normal scheme
         nx -= n_nos
@@ -1605,23 +1628,23 @@ def get_stretched_h(h_min, domain, nx, x0=0, x1=None, resp_domain=False):
             nx = nx_orig
             x0 = x0_orig
             x1 = None
-            h_min = h_min_orig
+            min_width = h_min_orig
 
     # Get stretching factor (a = 1+alpha).
-    if h_min == 0 or h_min > np.diff(domain)/nx:
-        # If h_min is bigger than the domain-extent divided by nx, no
+    if min_width == 0 or min_width > np.diff(domain)/nx:
+        # If min_width is bigger than the domain-extent divided by nx, no
         # stretching is required at all.
         alpha = 0
     else:
 
         # Wrap _get_dx into a minimization function to call with fsolve.
-        def find_alpha(alpha, h_min, args):
-            """Find alpha such that min(hx) = h_min."""
-            return min(get_hx(alpha, *args))/h_min-1
+        def find_alpha(alpha, min_width, args):
+            """Find alpha such that min(hx) = min_width."""
+            return min(get_hx(alpha, *args))/min_width-1
 
         # Search for best alpha, must be at least 0
         args = (domain, nx, x0)
-        alpha = max(0, optimize.fsolve(find_alpha, 0.02, (h_min, args)))
+        alpha = max(0, optimize.fsolve(find_alpha, 0.02, (min_width, args)))
 
     # With alpha get actual cell spacing with `resp_domain` to respect the
     # users decision.
@@ -1629,18 +1652,18 @@ def get_stretched_h(h_min, domain, nx, x0=0, x1=None, resp_domain=False):
 
     # Add the non-stretched center if x1 is provided
     if x1 is not None:
-        hx = np.r_[hx[: np.argmin(hx)], np.ones(n_nos)*h_min,
+        hx = np.r_[hx[: np.argmin(hx)], np.ones(n_nos)*min_width,
                    hx[np.argmin(hx):]]
 
-    # Print warning h_min could not be respected.
-    if abs(hx.min() - h_min) > 0.1:
+    # Print warning min_width could not be respected.
+    if abs(hx.min() - min_width) > 0.1:
         print(f"Warning :: Minimum cell width ({np.round(hx.min(), 2)} m) is "
-              "below `h_min`, because `nx` is too big for `domain`.")
+              "below `min_width`, because `nx` is too big for `domain`.")
 
     return hx
 
 
-def get_domain(x0=0, freq=1, rho=0.3, limits=None, min_width=None,
+def get_domain(x0=0, freq=1, res=0.3, limits=None, min_width=None,
                fact_min=0.2, fact_neg=5, fact_pos=None):
     r"""Get domain extent and minimum cell width as a function of skin depth.
 
@@ -1673,7 +1696,7 @@ def get_domain(x0=0, freq=1, rho=0.3, limits=None, min_width=None,
 
         Default is 1 Hz.
 
-    rho : float, optional
+    res : float, optional
         Resistivity (Ohm m) to calculate skin depth.
         Default is 0.3 Ohm m (sea water).
 
@@ -1713,7 +1736,7 @@ def get_domain(x0=0, freq=1, rho=0.3, limits=None, min_width=None,
         fact_pos = fact_neg
 
     # Calculate the skin depth.
-    skind = 503.3*np.sqrt(rho/abs(freq))
+    skind = 503.3*np.sqrt(res/abs(freq))
     if freq < 0:  # For Laplace-domain calculations.
         skind /= np.sqrt(2*np.pi)
 
@@ -1818,7 +1841,8 @@ def get_hx(alpha, domain, nx, x0, resp_domain=True):
 
                 # Note: this hx is equivalent as providing the following h
                 # to TensorMesh:
-                # h = [(h_min, nl-1, -a), (h_min, n_nos+1), (h_min, nr, a)]
+                # h = [(min_width, nl-1, -a), (min_width, n_nos+1),
+                #      (min_width, nr, a)]
 
     return hx